Add MatrixFloat type

+++ edge_list.h
@@ -16,7 +16,7 @@ private:
   int get_key(const Edge* s);
 public:
   const std::vector<Edge *> &get_edge_list() const;
-  EdgeList(const DiscreteArrayList &, int&);
+  EdgeList(const DiscreteArrayList &, size_t&);
   ~EdgeList();
 };
